Feature/test cases

+++ frontend/src/API/API.ts
@@ -4,9 +4,10 @@ import {getData} from 'src/LocalStorage';
 import './config';
 
 export const DELETE = async ({endpoint}: {endpoint: string}) => {
-  const accessToken = await getData('accessToken');
-
   try {
+    const accessToken = await getData('accessToken');
+    if(!accessToken) throw Error('no access token');
+
     const res = await axios({
       method: 'delete',
       url: endpoint,
@@ -16,7 +17,7 @@ export const DELETE = async ({endpoint}: {endpoint: string}) => {
     });
     return res;
   } catch (err) {
-    throw new Error(String(err));
+    throw Error(String(err));
   }
 };
 
@@ -27,10 +28,9 @@ export async function POST<T extends Object>({
   endpoint: string;
   reqBody: T;
 }) {
-  const accessToken = await getData('accessToken');
-
   try {
-    if (accessToken !== 'null') {
+    const accessToken = await getData('accessToken');
+    if (accessToken !== null || accessToken !== null) {
       const res = await axios({
         method: 'post',
         url: endpoint,
@@ -56,10 +56,9 @@ export async function POST<T extends Object>({
 }
 
 export const GET = async ({endpoint}: {endpoint: string}) => {
-  const accessToken = await getData('accessToken');
-
   try {
-    if (accessToken !== 'null') {
+    const accessToken = await getData('accessToken');
+    if (accessToken !== null || accessToken !== null) {
       const res = await axios({
         method: 'get',
         url: endpoint,
@@ -88,10 +87,9 @@ export async function PUT<T extends Object>({
   endpoint: string;
   reqBody: T;
 }) {
-  const accessToken = await getData('accessToken');
-
   try {
-    if (accessToken !== 'null') {
+    const accessToken = await getData('accessToken');
+    if (accessToken !== 'null' || accessToken !== null) {
       const res = await axios({
         method: 'put',
         url: endpoint,
